# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-Save Backend API
  description: |
    REST API for the Stellar-Save decentralised rotational savings platform.

    ## Authentication
    Most endpoints are public. Endpoints under `/api/user/*` require a Bearer JWT
    obtained from `POST /api/auth/verify`. Pass it as:
    ```
    Authorization: Bearer <token>
    ```

    ## Versioning
    The API is versioned via URL path prefix:
    - `/api/v1/` — stable, production-ready
    - `/api/v2/` — current development version (some routes return 501)

    Unversioned legacy paths (`/health`, `/recommendations`, etc.) are deprecated
    and will be removed on 2027-01-01.

    ## Rate Limiting
    - General endpoints: 100 req / 15 min per IP
    - Auth & admin endpoints: 10 req / 15 min per IP
    - Analytics write endpoints: stricter limits apply (see individual routes)
  version: 1.0.0
  contact:
    name: Stellar-Save
    url: https://github.com/Xoulomon/Stellar-Save
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.stellar-save.app
    description: Production
...
paths:
...
  /api/v1/stats/groups:
    get:
      tags: [Stats]
      summary: Platform-wide group statistics
      description: |
        Returns aggregated group statistics for the landing page.
        Sourced from the indexed ContractEvent database; cached 5 minutes in Redis.
      operationId: statsGroups
      responses:
        '200':
          description: Group overview statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsOverviewStats'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Recommendations ───────────────────────────────────────────────────────
  /api/v1/recommendations/{userId}:
    get:
      tags: [Recommendations]
      summary: Get group recommendations for a user (v1)
      description: |
        Uses A/B testing to select between content-based and collaborative
        filtering algorithms. Bucket A → content-based; Bucket B → collaborative.
      operationId: recommendationsV1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationsResponse'
  /api/v2/recommendations/{userId}:
    get:
      tags: [Recommendations]
      summary: Get group recommendations for a user (v2)
      description: Always uses collaborative filtering; returns richer metadata.
      operationId: recommendationsV2
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RecommendationsResponse'
                  - type: object
                    properties:
                      apiVersion:
                        type: string
                        example: v2
  /api/v1/preferences:
    post:
      tags: [Recommendations]
      summary: Set user group preferences
      description: Stores user preference data used by the recommendation engine.
      operationId: setPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferenceInput'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Search ────────────────────────────────────────────────────────────────
  /api/v1/search:
    get:
      tags: [Search]
      summary: Global full-text search
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/search/autocomplete:
    get:
      tags: [Search]
      summary: Autocomplete suggestions
      operationId: searchAutocomplete
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Autocomplete suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutocompleteResults'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Analytics ─────────────────────────────────────────────────────────────
  /api/v1/analytics/platform:
    get:
      tags: [Analytics]
      summary: Platform statistics for a date
      operationId: analyticsPlatform
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Target date (defaults to today)
      responses:
        '200':
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformStats'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/analytics/platform/trends:
    get:
      tags: [Analytics]
      summary: Platform statistics trends over a date range
      operationId: analyticsPlatformTrends
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Platform trends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformTrends'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/analytics/users/{userId}:
    get:
      tags: [Analytics]
      summary: User-specific analytics
      operationId: analyticsUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: User analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/analytics/groups/{groupId}:
    get:
      tags: [Analytics]
      summary: Group-specific analytics
      operationId: analyticsGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Group analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupStats'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/analytics/events:
    get:
      tags: [Analytics]
      summary: Analytics event statistics
      operationId: analyticsEventsGet
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Event statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalyticsEventStat'
    post:
      tags: [Analytics]
      summary: Record an analytics event
      operationId: analyticsEventsPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventInput'
      responses:
        '201':
          description: Event recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/analytics/cache/clear:
    post:
      tags: [Analytics]
      summary: Clear analytics cache
      operationId: analyticsCacheClear
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                  description: Redis key pattern to clear (default "*")
                  example: "analytics:platform:*"
      responses:
        '200':
          description: Cache cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ── Contract Events ───────────────────────────────────────────────────────
...
  /api/v1/events/stats:
    get:
      tags: [Events]
      summary: Contract event statistics
      operationId: eventsStats
      parameters:
        - name: contractId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Event statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventStats'

  # ── Export ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /api/auth/verify
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    GroupsOverviewStats:
      type: object
      properties:
        totalGroups:
          type: integer
        activeGroups:
          type: integer
        totalMembers:
          type: integer
        totalContributed:
          type: string
          description: Total XLM contributed across all groups (as string to avoid precision loss)
...